Added request for rental service and exception handler

+++ vhs/src/main/java/com/TrueNorth/vhs/service/RentalService.java
@@ -1,6 +1,7 @@
 package com.TrueNorth.vhs.service;
 
 import com.TrueNorth.vhs.entity.Rental;
+import com.TrueNorth.vhs.exception.ResourceNotFoundException;
 import com.TrueNorth.vhs.request.VhsRequest;
 import com.TrueNorth.vhs.service.VhsService;
 import com.TrueNorth.vhs.entity.User;
@@ -28,11 +29,12 @@ public class RentalService {
 
     private final RentalRepository rentalRepository;
     VhsService vhsService;
-
+    private final VhsRepository vhsRepository;
 
     @Autowired
-    public RentalService(RentalRepository rentalRepository){
+    public RentalService(RentalRepository rentalRepository,VhsRepository vhsRepository){
         this.rentalRepository=rentalRepository;
+        this.vhsRepository=vhsRepository;
     }
 
 
@@ -47,29 +49,65 @@ public class RentalService {
         Rental rental = new Rental();
         rental.setUser(rentalRequest.getUser_id());
         rental.setVhs(rentalRequest.getVhs_id());
-        rentalRepository.updateVhs(rental.getVhs().getVhsId());
-        return rentalRepository.save(rental);
+        Optional<Vhs> getVhs = vhsRepository.findById(rental.getVhs().getVhsId());
+        Vhs vhs =getVhs.get();
+        if(vhs.isIs_rental() == false) {
+            rentalRepository.updateVhs(rental.getVhs().getVhsId());
+            return rentalRepository.save(rental);
+
+        }
+        else{
+            throw new ResourceNotFoundException("Vhs is already rented");
+        }
 
     }
 
 
-    public void deleteRentalById(Long id){
+    public void deleteRentalById(Long id) {
 
-      Optional<Rental> getRental = rentalRepository.findById(id);
+        Optional<Rental> getRental = Optional.ofNullable(rentalRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Rental not found with id:" + id)));
+
       Rental rental =getRental.get();
-      rentalRepository.setVhsToFalse(rental.getVhs().getVhsId());
-      rentalRepository.deleteById(id);
+
+      if(checkRentedDay(rental.getReturned(),LocalDate.now()) == 1) {
+          rentalRepository.setVhsToFalse(rental.getVhs().getVhsId());
+          rentalRepository.deleteById(id);
+      }
+      else
+          throw new ResourceNotFoundException("Too late");
+
 
     }
+
+
+    public int checkRentedDay(LocalDate d1 , LocalDate d2){
+        int result = d1.compareTo(d2);
+        if(result == 0){
+            return 1;
+        }
+        if(result > 0){
+            return 1;
+        }
+        else
+            return 0;
+    }
+
 
     public ResponseEntity<Rental> getRentalById(Long id){
-        Optional<Rental> getRental = rentalRepository.findById(id);
+
+        Optional<Rental> getRental = Optional.ofNullable(rentalRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Rental not found with id:" + id)));
         Rental rental= getRental.get();
         return new ResponseEntity<>(rental, HttpStatus.OK);
     }
 
+
+
+
     public Rental updateRental(RentalRequest rentalRequest,Long id){
-        Optional<Rental> getRental = rentalRepository.findById(id);
+        Optional<Rental> getRental = Optional.ofNullable(rentalRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Rental not found with id:" + id)));
         Rental rental=getRental.get();
 
         rental.setReturned(rentalRequest.getReturned());
